main/requirements.txt
Augmentor==0.2.12
einops==0.6.0
matplotlib==3.7.0
numpy==1.23.1
opencv_python==4.7.0.68
pandas==1.5.3
Pillow==9.4.0
torch==1.13.1
torchmetrics==0.11.1
torchvision==0.14.1
tqdm==4.64.1
